+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-For a full diff see [`0.3.1...main`][0.3.1...main].
+For a full diff see [`0.4.0...main`][0.4.0...main].
+
+## [`0.4.0`][0.4.0] - 2025-03-21
+
+For a full diff see [`0.3.1...0.4.0`][0.3.1...0.4.0].
 
 ### Changed
 
@@ -41,6 +45,8 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 
 [0.3.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.3.0
 [0.3.1]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.3.1
+[0.4.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.4.0
 [0.2.2...0.3.0]: https://github.com/weph/obsidian-bookshelf/compare/0.2.2...0.3.0
 [0.3.0...0.3.1]: https://github.com/weph/obsidian-bookshelf/compare/0.3.0...0.3.1
-[0.3.1...main]: https://github.com/weph/obsidian-bookshelf/compare/0.3.1...main
+[0.3.1...0.4.0]: https://github.com/weph/obsidian-bookshelf/compare/0.3.1...0.4.0
+[0.4.0...main]: https://github.com/weph/obsidian-bookshelf/compare/0.4.0...main